move utils from components to original utils folder
