+++ src/components/App.js
@@ -3,7 +3,7 @@ import { firestoreConnect } from 'react-redux-firebase';
 import { compose } from 'redux';
 import '../App.css';
 import { Sidebar, Board, ActionFooter } from './index';
-import { doc } from './utils';
+import { doc } from '../utils';
 // import { db } from '../store';
 // import { init, collections, setupLogic } from '../utils';
