+++ src/components/PlayerHand.js
@@ -3,7 +3,7 @@ import { Image, Container } from 'semantic-ui-react';
 import { connect } from 'react-redux';
 import { compose } from 'redux';
 import { firestoreConnect, isLoaded } from 'react-redux-firebase';
-import { doc } from './utils';
+import { doc } from '../utils';
 
 const PlayerHand = ({ game, currentHand, eventCards } ) => {
   return (
