--- src/components/utils/index.js
@@ -1,2 +1,0 @@
-export * from './actionFooter';
-export { doc } from './currentDoc';
