+++ src/utils/components/choosePlayerModal.js
@@ -1,0 +1,9 @@
+export const playerImage = {
+  'Contingency Planner': 'assets/images/cont_planner.png',
+  Dispatcher: 'assets/images/dispatcher.png',
+  Medic: 'assets/images/medic.png',
+  'Operations Expert': 'assets/images/ops_expert.png',
+  'Quarantine Specialist': 'assets/images/quar_spec.png',
+  Researcher: 'assets/images/researcher.png',
+  Scientist: 'assets/images/scientist.png'
+};
