+++ src/utils/components/index.js
@@ -1,0 +1,4 @@
+export * from './actionFooter';
+export * from './currentDoc';
+export * from './choosePlayerModal';
+export * from './cityLines';
