+++ src/utils/index.js
@@ -1,2 +1,3 @@
 export * from './cards';
 export * from './gameLogic';
+export * from './components';
